+++ components/rightLayout/Price.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
-import { Button } from '../../stories/Button';
+import { Button } from '../common/Button';
 import IconPrice from './../../assets/images/icon_price.svg';
 const Price = () => {
   return (
