--- stories/Header/header.module.css
@@ -1,76 +1,0 @@
-.wrapper {
-  height: 55px;
-  width: 100%;
-  background-color: #49515B;
-}
-
-
-.h1 {
-  font-weight: 700;
-  line-height: 1;
-  margin: 6px 0 6px 10px;
-  display: inline-block;
-  vertical-align: top;
-  font-size: 16px;
-  color: #ECECEC;
-}
-
-.innerWrapper{
-  max-width: 1600px;
-  height: 55px;
-  padding : 0 80px;
-  align-items: center;
-  margin: 0 auto;
-  display: flex;
-  justify-content: space-between;
-  position: relative;
-}
-
-.welcome {
-  color: #333;
-  font-size: 14px;
-  margin-right: 10px;
-}
-.floatingSearch{
-  height: 34px;
-  width:340px;
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-}
-.input {
-width: 100%;
-height: 34px;
-border-radius: 20px;
-padding: 0 15px;
-}
-.input::placeholder {
-  text-align: center;
-  transition: 0.3;
-}
-.floatingSearchicon{
-  display: block;
-  width: 18px;
-  height: 18px;
-  background-image: url('./../../assets/images/icon_search.png');
-  background-position: 100% 100%;
-  background-size: cover;
-  position: absolute;
-  top: 50%;
-  left: calc(50% - 40px);
-  transform: translate(-50%, -50%);
-}
-.floatingSearchiconactive{
-  left: 30px;
-}
-.spanSearch{
-  color:#8B8B8B;
-  font-size: 14px;
-  width: 100%;
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  text-align: center;
-}
